Refactor inactivity warning into global context
